requirements.txt
django
django-rest-framework
requests
djangorestframework-simplejwt
corsheaders
drf-yasg
pytest-django
drf-yasg
